0.2.0

+++ lib/app/database/repository/pill_repository.dart
@@ -1,12 +1,15 @@
 import '../../models/pill_model.dart';
+import '../../utils/result.dart';
 import '../app_database.dart';
 import '../dao/pill_dao.dart';
+import '../dao/plan_dao.dart';
 
 class PillRepository {
   final AppDatabase _db;
   final PillDao _pillDao;
+  final PlanDao _planDao;
 
-  PillRepository(this._db, this._pillDao);
+  PillRepository(this._db, this._pillDao, this._planDao);
 
   Future<bool> isEmpty() {
     return _pillDao.isEmpty();
@@ -24,7 +27,7 @@ class PillRepository {
     });
   }
 
-  Future<int> addPill(PillModel pill) {
+  Future<Result<int>> addPill(PillModel pill) {
     return _db.transaction(() {
       return _pillDao.addPill(pill);
     });
@@ -36,9 +39,12 @@ class PillRepository {
     });
   }
 
-  Future<void> deletePill(int id) {
-    return _db.transaction(() {
-      return _pillDao.deletePill(id);
+  Future<Result<void>> deletePill(int id) {
+    return _db.transaction(() async {
+      final r = await _planDao.hasPlanByPill(id);
+      if (r) return Result.failure('此药物有计划，请先删除计划');
+      await _pillDao.deletePill(id);
+      return Result.success(null);
     });
   }
 }
